backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.5
python-dotenv==1.0.1
openai==1.13.3
httpx==0.27.0
